Revert "Merge branch 'trunk' into feature/11925-app-settings-cleanup"
This reverts commit 1e09f4e38851f481fce4cf860ec91aa34ce52472.

+++ WordPress/Classes/Utility/AccountHelper.swift
@@ -84,9 +84,6 @@ import Foundation
 
         service.removeDefaultWordPressComAccount()
 
-        // Delete saved dashboard states
-        BlogDashboardState.resetAllStates()
-
         // Delete local notification on logout
         PushNotificationsManager.shared.deletePendingLocalNotifications()
 
@@ -97,6 +94,8 @@ import Foundation
         StatsDataHelper.clearWidgetsData()
 
         // Delete donated user activities (e.g., for Siri Shortcuts)
-        NSUserActivity.deleteAllSavedUserActivities {}
+        if #available(iOS 12.0, *) {
+            NSUserActivity.deleteAllSavedUserActivities {}
+        }
     }
 }
